requirements.txt
streamlit==1.8.1
streamlit-option-menu == 0.3.6
streamlit-camera-input-live == 0.2.0